+++ src/components/event/components/CardEvent.jsx
@@ -1,5 +1,3 @@
-"use client"
-
 import { Button } from "@nextui-org/react"
 import Image from "next/image"
 import Sample from "../../../../public/assets/dummy.png"
